requirements.txt
asyncpg>=0.29.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
email-validator>=2.2.0
PyJWT>=2.9.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
